Updated PORT to .env and fixed some known issues

+++ backend/controllers/product.controllers.js
@@ -1,4 +1,4 @@
-import Product from './models/product.model.js';
+import Product from '../models/product.model.js';
 import mongoose from 'mongoose';
 
 export const getProducts = async (req,res)=>{
@@ -44,11 +44,14 @@ export const updateProducts = async (req,res)=>{
 export const deleteProducts = async (req,res)=>{
     const {id} = req.params;
     console.log("ID : ", id);
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(400).json({success:false, message:"Invalid Product ID"});
+    }
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({success:true, message:"Product deleted successfully"});
     } catch (error) {
-        res.status(404).json({success:false, message:"Product Not Found"});
+        res.status(500).json({success:false, message:"Product Not Found Server Error"});
     }
 
 }
